+++ R/mod_codebook_utils_codebook.R
@@ -170,8 +170,10 @@ list_db_codes <- function(pool, project_id, user) {
     ## To pass R CMD check and define DB variables as global variables for the function https://www.r-bloggers.com/2019/08/no-visible-binding-for-global-variable/
     code_id <- code_name <- code_description <- code_color <- NULL
 
+    active_project <- as.integer(unname(project_id))
+
     project_codes <- dplyr::tbl(pool, "codes") %>%
-        dplyr::filter(.data$project_id == as.integer(.env$project_id)) %>%
+        dplyr::filter(.data$project_id == active_project) %>%
         dplyr::select(
             code_id,
             code_name,
@@ -334,6 +336,36 @@ convert_colour_to_hex <- function(color) {
     )
 }
 
+# prepare data.frame with codes and categories to export
+get_codebook_export_table <- function(glob) {
+    active_project <- as.integer(unname(glob$active_project))
+
+    categories <- dplyr::tbl(glob$pool, "categories") %>%
+        dplyr::filter(project_id == active_project) %>%
+        dplyr::select(category_id, category_name, category_description) %>%
+        dplyr::collect() %>%
+        dplyr::mutate(
+            category_title = dplyr::if_else(
+                !is.na(category_description) & category_description != "",
+                paste0(category_name, " (", category_description, ")"),
+                category_name
+            )
+        ) %>%
+        dplyr::select(-c(category_name, category_description))
+
+    categories_map <- dplyr::tbl(glob$pool, "categories_codes_map") %>%
+        dplyr::collect() %>%
+        dplyr::inner_join(categories, by = "category_id")
+
+    dplyr::left_join(glob$codebook, categories_map, by = "code_id") %>%
+        dplyr::group_by(code_id, code_name, code_description, code_color) %>%
+        dplyr::summarise(
+            categories = paste0(category_title, collapse = " | ")
+        ) %>%
+        dplyr::mutate(
+            code_color = purrr::map_chr(code_color, convert_colour_to_hex)
+        )
+}
 
 # Edit codes ------
 
@@ -346,6 +378,8 @@ edit_db_codes <- function(
     edit_code_description,
     edit_code_color
 ) {
+    active_project <- as.integer(unname(active_project))
+
     update_code_sql <- glue::glue_sql(
         "UPDATE codes
                  SET code_name = {edit_code_name}, code_description = {edit_code_description}, code_color = {edit_code_color}